requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.7.0
requests